requirements.txt
streamlit>=1.30.0
numpy<2.0.0
Pillow>=10.0.0
ultralytics>=8.1.0
opencv-python-headless>=4.8.0
